+++ src/ScottPlot5/ScottPlot5/Plot.cs
@@ -1,6 +1,7 @@
 ﻿using ScottPlot.AxisPanels;
 using ScottPlot.Control;
 using ScottPlot.Legends;
+using ScottPlot.Primitives;
 using ScottPlot.Rendering;
 using ScottPlot.Stylers;
 
@@ -13,11 +14,9 @@ public class Plot : IDisposable
     public RenderManager RenderManager { get; }
     public RenderDetails LastRender => RenderManager.LastRender;
     public LayoutManager Layout { get; private set; }
-    public Color FigureBackground { get; set; } = Colors.White;
-    public Color DataBackground { get; set; } = Colors.White;
-    public SKBitmap? DataBackgroundImage { get; set; } = null;
-    public ImageScalingStyle DataBackgroundScalingStyle { get; set; } = ImageScalingStyle.FillRetainAspect;
-    public Color DataBackgroundImageColor { get; set; } = Colors.White;
+
+    public BackgroundStyle FigureBackground = new();
+    public BackgroundStyle DataBackground = new();
 
     public IZoomRectangle ZoomRectangle { get; set; } = new StandardZoomRectangle();
     public float ScaleFactor { get; set; } = 1.0f;
@@ -42,7 +41,8 @@ public class Plot : IDisposable
 
     public void Dispose()
     {
-        DataBackgroundImage?.Dispose();
+        DataBackground?.Dispose();
+        FigureBackground?.Dispose();
         PlottableList.Clear();
         Axes.Clear();
     }