+++ CHANGELOG.md
@@ -17,6 +17,7 @@ _Not yet on NuGet..._
 * SignalXY: Improve support for displaying data on inverted axes (#3396, #3400) @BrianAtZetica
 * Axes: Improved support for ticks and labels on inverted axes (#3401, #3397) @BrianAtZetica
 * Plot: Added `Remove()` overloads for Axes, Panels, and Grids (#3402, #3360) @Excustic, @redrabbit007, @csbebetter, @xichaoqiang
+* Plot: `Plot.FigureBackground` is now `Plot.FigureBackground.Color` (and same with `DataBackground`)
 
 ## ScottPlot 5.0.21
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-01-28_